Revert "Upgrade to 2.39.5"
This reverts commit cc3da95bc5526397004e0dc326ff4aa07771080f.

Why? Because this won't build successfully, because I didn't test it
locally since there were only a couple days since .4, but there's a
soname bump because we removed some APIs.

And we're doing a mass rebuild right now, and it's not ideal for that to
pick up a known-bad commit.